refactor : improved the auth layout and auth pages, used in page form, removed form sepration from the page
- Improved folder option for auth pages
- used layout and auth card for dynmic headers
- removed form pages instead added those forms in page directory itself
